File writer (#3)
Decided to just do a pull request now that the reset button and data output method have been implemented and tested.

Tried to implement automation to play multiple games in a row, but it didn't work at all due to playGame() being an asynchronous recursive function, so I commented it out for the time being. I'll take another crack at it again on Monday, but it looks like it's going to be a hassle.

For now, I think it's best to just merge in the reset and output functionality.